chore: try remove some SSR-related workarounds (#27596)
