+++ packages/devextreme/js/__internal/core/license/license_validation.ts
@@ -13,8 +13,7 @@ import { INTERNAL_USAGE_ID, PUBLIC_KEY } from './key';
 import { pad } from './pkcs1';
 import { compareSignatures } from './rsa_bigint';
 import { sha1 } from './sha1';
-import type { CustomTrialPanelStyles } from './trial_panel';
-import { registerCustomComponents, renderTrialPanel } from './trial_panel';
+import { showTrialPanel } from './trial_panel';
 import type {
   License,
   LicenseCheckParams,
@@ -155,22 +154,6 @@ function getLicenseCheckParams({
     return { preview, error: undefined };
   } catch {
     return { preview, error: 'W0021' };
-  }
-}
-
-export function showTrialPanel(
-  buyNowUrl: string,
-  version: string,
-  customStyles?: CustomTrialPanelStyles,
-): void {
-  if (typeof customElements !== 'undefined') {
-    renderTrialPanel(buyNowUrl, version, customStyles);
-  }
-}
-
-export function registerTrialPanelComponents(customStyles?: CustomTrialPanelStyles): void {
-  if (typeof customElements !== 'undefined') {
-    registerCustomComponents(customStyles);
   }
 }
