+++ packages/devextreme/js/__internal/core/license/license_validation_internal.ts
@@ -1,19 +1,10 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import type { CustomTrialPanelStyles } from './trial_panel';
 import type { Token } from './types';
 
 // @ts-expect-error - only for internal usage
 export function parseLicenseKey(encodedKey: string | undefined): Token {}
 
 export function validateLicense(licenseKey: string, version?: string): void {}
-
-export function showTrialPanel(
-  buyNowUrl: string,
-  version: string,
-  customStyles?: CustomTrialPanelStyles,
-): void {}
-
-export function registerTrialPanelComponents(customStyles?: CustomTrialPanelStyles): void {}
 
 // @ts-expect-error - only for internal usage
 export function peekValidationPerformed(): boolean {}
