+++ packages/devextreme/js/core/trial_panel.js
@@ -1,4 +1,4 @@
-import { registerTrialPanelComponents, showTrialPanel } from '../__internal/core/license/license_validation';
+import { registerTrialPanelComponents, showTrialPanel } from '../__internal/core/license/trial_panel';
 
 export default {
     registerTrialPanelComponents,
